split the source
